Move validation around, remove FilterClause subtypes (base class sufficient)
